Two files stop being unreadable, and three dead metrics go
## `authorization/service.py`: 1926 lines -> four files

`AuthorizationDataService` and `Authorizer` shared nothing but an
`AsyncSession`. Nothing in `Authorizer` calls the data service; the dependency
runs one way, which is why the import points from `service` to `authorizer` and
not back.

    service.py           1926  ->  1067   role CRUD, context building
    authorizer.py                    486   the decision engine
    grant_resolution.py              347   which grants apply, and what they confer
    hydration.py                     274   turning a reference into facts

The two extracted halves are mixins because they call each other and read
`self.session`; `Authorizer` composes both, which is the shape
`SurfaceConsentMixin` already uses here.

**Every `app/modules/*` import in the authorization path is now in
`hydration.py`.** That does not reduce `core_module_imports` -- it is still
`app/core` -- and the PR body says so. What it buys is that the later inversion
becomes a change to one 274-line file instead of to the decision engine.

## `app.py`: 999 -> 496

    app.py         999  ->  496
    middleware.py            344   three self-contained ASGI classes
    health.py                241   liveness, readiness, capabilities

`oversized_files` **20 -> 19**: `app.py` is the one file that crossed under the
600-line ceiling, which is the only way that count moves.

The probes become an `APIRouter`, registered before the module routers so a
module router raising while it is included cannot take readiness down with it.
`health_ready` read `app.state.embedded_worker` off a closure; it takes a
`Request` and reads `request.app.state`, the same object.

## Three metrics deleted

`composition_deep_imports`, `module_composition_imports` and
`induced_module_cycles` all measured `app/composition`, which was emptied and
deleted. They did their job. `_inline_composition` -- which existed to report
what the graph would look like once the hop was gone -- had nothing left to
inline, so `induced_module_cycles` was `module_cycles` computed twice.

Gone rather than kept at zero: a metric that cannot move is not a ratchet, it is
a line a reader has to work out is dead. `design.md`'s DES-02 and DES-03 are
recorded as done with the numbers they reached, and DES-01/04/09 are re-measured
(34 -> 23, 40 -> 21, 23 -> 20). The rest were not re-measured and now say so,
because a stale number is worth less than none.

## What the split shook loose

Every one of these failed loudly, which is the good version:

- Tests patching `service_module.has_session_approval`, `app.app.RequestObserverMiddleware`
  and four probe collaborators -- all names that moved. `monkeypatch` raises on a
  missing attribute, so each said exactly what had gone.
- `test_severity_policy` rebinds a module's logger to see DEBUG output. The
  request observer carries its own logger now, so rebinding `app.app`'s alone
  stopped reaching it.
- `test_api_docs_exposure` asserted `/livez` was in `app.routes`. FastAPI 0.141
  wraps every `include_router` in an `_IncludedRouter`, so that flat scan had
  only ever seen the five routes registered directly on the app -- it was blind
  to all sixty-three module routers, and moving the probes into one revealed it.
  It walks properly now, except for the `/docs` assertions, which stay flat
  deliberately: FastAPI registers those on the app, and a deep walk finds a
  module's own `/docs` route and fails for the wrong reason.
- One `except Exception` moved file, so the swallowed-errors baseline reports it
  as one gone and one new. Same handler, new key; the count is unchanged.
- `middleware.py` arrived with a `scope: dict`, which grew `untyped_escapes` by
  a file. Typed as `starlette.types.Scope` rather than baselined.

`make test-unit` 7483 passed; `SHARD=datastore` 269 passed; `make quality` green.
