+++ docs/engineering/design.md
@@ -53,6 +53,10 @@ the recipe:
 
 ## Boundaries
 
+> The `Today` figures under DES-01, DES-02/03, DES-04, DES-05 and DES-09 were
+> re-measured in Sep 2026. The others were not, and a stale number is worth less
+> than none — re-run each rule's own `Check` before quoting it.
+
 ### DES-01 — a module reaches another module only through `contracts` or a domain event
 
 Not its services, not its repositories, not its ORM models, not its
@@ -62,48 +66,41 @@ Not its services, not its repositories, not its ORM models, not its
 making. It survives refactors by preventing them.
 
 *Check:* `uv run python scripts/check_architecture.py` (`forbidden_imports`)
-*Today:* **34** — ratchet
+*Today:* **23** over 9 module pairs — ratchet
 
-### DES-02 — the composition root composes; it does not re-export
+### DES-02, DES-03 — the composition root: **done**
 
-`app/composition/` is being emptied, and DES-03 below is the rule that replaces
-it: a capability belongs to the module that provides it, published from that
-module's `contracts/`. A file there whose whole body is `from … import X` with an
-`__all__` is coupling with a nicer address — `app/composition/surface_agent.py`
-is the last one, and its docstring says what it is waiting on.
+Both rules governed `app/composition/`, a shared middle layer that thirteen of
+the fifteen modules depended on. The directory has been emptied and deleted, so
+neither rule has anything left to govern: a capability now belongs to the module
+that provides it and is published from that module's `contracts/`, which is
+DES-01.
 
-*Check:* `check_architecture.py` (`composition_deep_imports`)
-*Today:* **32** composition→module imports reach past `contracts` — ratchet
+The three metrics that measured it — `composition_deep_imports`,
+`module_composition_imports` and `induced_module_cycles` — are gone from
+`check_architecture.py` too. The last of those existed to report what the graph
+would look like once the hop was deleted; with the hop gone it had nothing to
+inline and was `module_cycles` computed twice.
 
-### DES-03 — a module must not import the composition root
-
-Dependencies point inward. When a module imports `app.composition`, the root is
-no longer a root; it is a shared middle layer that every module is coupled to.
-
-*Check:* `make architecture` — `module_composition_imports` in the baseline.
-*Today:* **25** — ratchet
-
-Until this rule had a check, the number was a `grep` in this document and nothing
-failed when it rose. Worse, the cycles those edges carried were invisible too:
-`module_cycles` reads **0** because a file under `app/composition` is excluded
-from the dependency graph, so a cycle with a hop through the root is not a cycle
-as far as the gate is concerned. `induced_module_cycles` inlines that hop.
-It reports **one component of 13 of the 15 modules**, and that is the number
-this rule is really about — deleting the root without breaking those cycles
-first would turn the whole backend into one knot.
+The numbers are kept here as a record of what the rules were worth: 32
+composition→module imports past `contracts`, 25 modules reaching the root, and
+one knot containing 13 of the 15 modules. All three are now zero, and
+`module_cycles` reads **0** because there are none rather than because a
+directory was excluded from the graph.
 
 ### DES-04 — `app/core/` must not import `app/modules/`
 
 Core is what modules are built on. The one exception is
 `app/core/registry/installed.py`, whose job is naming them.
 
-Today the central authorization service imports eight modules' ORM classes. The
-gate could not see any of it until recently: it scanned `app/modules/` only,
-which is how the count reached 42 without anyone deciding on it.
+The central authorization service still imports four modules' ORM classes, and
+the resource tables and name registry account for most of the rest. The gate
+could not see any of it until recently: it scanned `app/modules/` only, which is
+how the count reached 42 without anyone deciding on it.
 
 *Check:* `check_architecture.py` (`core_module_imports`), which exempts
 `app/core/registry/installed.py`
-*Today:* **40** — ratchet
+*Today:* **21** over 9 modules — ratchet
 
 ### DES-05 — layers point one way
 
@@ -195,9 +192,8 @@ find. The current gate applies the rule to `app/modules/` only, which is how
 `app/composition/analytics_consumer.py` reached 975 lines unnoticed.
 
 *Check:* `check_architecture.py` (`oversized_files`)
-*Today:* **23** — 14 under `app/modules/` and 9 that only became visible when the
-gate was widened past it, including `app/composition/analytics_consumer.py` at
-975 lines — ratchet
+*Today:* **20** — 13 under `app/modules/` and 7 under `app/core/`, which only
+became visible when the gate was widened past `app/modules/` — ratchet
 
 ### DES-10 — split a service by use case, not by layer
