+++ lemma-backend/app/app.py
@@ -1,12 +1,8 @@
 import asyncio
-import re
-import time
-import uuid
 from collections.abc import Mapping, Sequence
 from contextlib import AsyncExitStack, asynccontextmanager
 from fastapi import Depends, FastAPI
 from opentelemetry import metrics
-from fastapi.responses import JSONResponse
 from fastapi.openapi.utils import get_openapi
 from scalar_fastapi import get_scalar_api_reference
 from starlette.middleware.cors import CORSMiddleware
@@ -17,10 +13,14 @@ from app.version import API_VERSION
 from app.core.api.session_cookie_scope import RefreshCookieScopeMiddleware
 from app.core.api.exception_handlers import register_exception_handlers
 from app.core.api.streaming_multipart import install_streaming_multipart_openapi
-from app.core.domain.errors import PayloadTooLargeError
 from app.core.config import settings
+from app.health import router as health_router
+from app.middleware import (
+    RequestBodyLimitMiddleware,
+    RequestIdMiddleware,
+    TrailingSlashMiddleware,
+)
 from app.core.cors import get_allowed_cors_origin_regex, get_allowed_cors_origins
-from app.core.origin import origin_for_path, origin_scope, resolve_client_identity
 from app.core.infrastructure.events.message_bus import (
     close_message_bus,
     get_message_bus,
@@ -47,12 +47,7 @@ from app.core.observability.telemetry import (
     instrument_fastapi_app,
     shutdown_telemetry,
 )
-from app.core.infrastructure.db.migration_state import schema_migration_state
-from app.core.observability.dependency_incident import DependencyIncident
-from app.core.observability import readiness
-from app.core.observability.worker_liveness import worker_readiness_state
-from app.core.security import supertokens_core_reachable
-from app.sandbox_health import record_sandbox_probe, sandbox_capability
+from app.sandbox_health import record_sandbox_probe
 from app.core.infrastructure.channels.channel_service import channel_service
 
 from app.modules.apps.api.host_routing import AppHostRoutingMiddleware
@@ -60,11 +55,9 @@ from app.core.registry.assembly import enter_api_lifespans, include_module_route
 from app.core.registry.installed import OSS_MODULES
 from app.auth_app import get_auth_app
 from app.mcp_server import get_agent_mcp_app, get_pod_mcp_app
-from app.core.infrastructure.db.session import database_reachable, get_engine
+from app.core.infrastructure.db.session import get_engine
 from app.core.request_context import (
-    bind_request_context,
     create_background_task,
-    create_inherited_task,
 )
 
 logger = get_logger(__name__)
@@ -297,308 +290,6 @@ async def lifespan(app: FastAPI):
 #: outcome (a 404, a sub-app mount, an error before routing), not an error --
 #: but it is also not an identity, which is why the raw path rides along on the
 #: signals that matter.
-_UNMATCHED_ROUTE = "unmatched"
-
-
-class TrailingSlashMiddleware:
-    """Treat ``/things/`` as ``/things`` so a stray slash is not a 404.
-
-    Mutates the scope in place rather than copying it, for the same reason
-    documented in ``apps/api/host_routing.py``: the router records the matched
-    route by writing ``scope["route"]``, and :class:`RequestObserverMiddleware`
-    reads it from *outside* this middleware. With a copy the router wrote to an
-    object the observer never saw, so every request whose path ended in a slash
-    was logged as ``route: "unmatched"`` regardless of what it actually matched
-    or returned -- silently seeding the bucket that a fixed-cost class of slow
-    "unmatched" 404s was being investigated in.
-    """
-
-    def __init__(self, app):
-        self.app = app
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] != "http":
-            await self.app(scope, receive, send)
-            return
-
-        path = scope.get("path", "")
-        if path != "/" and path.endswith("/"):
-            scope["path"] = path.rstrip("/")
-
-        await self.app(scope, receive, send)
-
-
-class RequestObserverMiddleware:
-    """Bind HTTP correlation, emit bounded terminal signals, and record metrics."""
-
-    HEADER = b"x-request-id"
-    # How the work arrived, per docs/design/product-analytics.md. Resolved once
-    # here so every downstream emit reads it from context rather than guessing.
-    CLIENT_HEADER = b"x-lemma-client"
-    REQUEST_ID_RE = re.compile(r"^[A-Za-z0-9._:-]{1,128}$")
-    SLOW_SECONDS = 2.0
-    QUIET_PATHS = frozenset(
-        {"/health", "/health/live", "/health/ready", "/health/capabilities", "/livez"}
-    )
-    # Routes that are supposed to take a long time. A long poll answers when it
-    # has news or when its hold expires, so a slow one is the design working:
-    # every completed idle poll logged a warning, and on one local stack 311 of
-    # 314 slow-request warnings were healthy 25-second polls. A warning that
-    # fires on the normal case is not a signal, and it buried the three that
-    # meant something.
-    HELD_ROUTES = frozenset({"/agent-host/poll"})
-
-    def __init__(self, app):
-        self.app = app
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] != "http":
-            await self.app(scope, receive, send)
-            return
-
-        headers = list(scope.get("headers") or [])
-        existing = next((v for k, v in headers if k.lower() == self.HEADER), None)
-        inbound = existing.decode("latin-1") if existing is not None else ""
-        if self.REQUEST_ID_RE.fullmatch(inbound):
-            request_id = inbound
-        else:
-            request_id = uuid.uuid4().hex
-
-        correlation_id = uuid.uuid7()
-        scope = dict(scope)
-        scope["headers"] = [
-            (key, value) for key, value in headers if key.lower() != self.HEADER
-        ] + [(self.HEADER, request_id.encode("ascii"))]
-        scope.setdefault("state", {})
-
-        started_at = time.perf_counter()
-        response_started_at: float | None = None
-        status_code = 500
-        content_type = ""
-
-        async def send_with_request_id(message):
-            nonlocal response_started_at, status_code, content_type
-            if message["type"] == "http.response.start":
-                response_started_at = time.perf_counter()
-                status_code = int(message.get("status", 500))
-                raw_headers = [
-                    (key, value)
-                    for key, value in list(message.get("headers") or [])
-                    if key.lower() != self.HEADER
-                ]
-                content_type = next(
-                    (
-                        value.decode("latin-1").lower()
-                        for key, value in raw_headers
-                        if key.lower() == b"content-type"
-                    ),
-                    "",
-                )
-                raw_headers.append((self.HEADER, request_id.encode("ascii")))
-                message = {**message, "headers": raw_headers}
-            await send(message)
-
-        client_header = next(
-            (v for k, v in headers if k.lower() == self.CLIENT_HEADER), None
-        )
-        # The mount point wins over the header where the route itself settles
-        # the question: an MCP caller sends no Lemma client header.
-        resolved_origin = origin_for_path(scope.get("path") or "") or (
-            resolve_client_identity(
-                client_header.decode("latin-1", "replace") if client_header else None
-            ).origin
-        )
-
-        caught: Exception | None = None
-        cancelled = False
-        with (
-            bind_request_context(request_id=request_id, correlation_id=correlation_id),
-            origin_scope(resolved_origin),
-        ):
-            try:
-                await self.app(scope, receive, send_with_request_id)
-            except asyncio.CancelledError:
-                cancelled = True
-                raise
-            except Exception as exc:
-                caught = exc
-                raise
-            finally:
-                finished_at = time.perf_counter()
-                duration_ms = round((finished_at - started_at) * 1000, 1)
-                route = self._route_template(scope)
-                attributes = {
-                    "http.request.method": str(scope.get("method", "UNKNOWN")),
-                    "http.route": route,
-                    # The exact code, not the class. The FastAPI instrumentation's
-                    # own histogram records exact codes but no route, and this
-                    # counter records the route -- matching the vocabularies is
-                    # what lets a dashboard join them into per-route error rate.
-                    # Cardinality is bounded by the codes we actually return.
-                    "http.response.status_code": status_code,
-                }
-                http_request_count.add(1, attributes)
-                http_request_duration.record(duration_ms, attributes)
-
-                if str(scope.get("path", "")) in self.QUIET_PATHS:
-                    continue_logging = False
-                else:
-                    continue_logging = True
-                if continue_logging and not cancelled:
-                    state = scope.get("state") or {}
-                    recorded = state.get("lemma_exception")
-                    failure = caught or recorded
-                    fields = {
-                        "method": str(scope.get("method", "UNKNOWN")),
-                        "route": route,
-                        "status_code": status_code,
-                        "duration_ms": duration_ms,
-                    }
-                    # "unmatched" names no request, so a slow or failing one in
-                    # that bucket cannot be investigated at all: two separate
-                    # passes at a fixed-cost class of slow `unmatched` 404s
-                    # failed for exactly this reason. Populated only for that
-                    # bucket -- a real route template is already the identity,
-                    # and raw paths are unbounded.
-                    fields["path"] = (
-                        str(scope.get("path", ""))[:120]
-                        if route == _UNMATCHED_ROUTE
-                        else ""
-                    )
-                    if status_code >= 500 or caught is not None:
-                        fields["error_type"] = state.get(
-                            "lemma_error_type",
-                            type(failure).__name__ if failure else "HTTPError",
-                        )
-                        fields["error_code"] = state.get(
-                            "lemma_error_code", "INTERNAL_ERROR"
-                        )
-                        exc_info = (
-                            (type(failure), failure, failure.__traceback__)
-                            if isinstance(failure, BaseException)
-                            else None
-                        )
-                        logger.error(
-                            "http.request.failed",
-                            method=fields["method"],
-                            route=fields["route"],
-                            status_code=fields["status_code"],
-                            duration_ms=fields["duration_ms"],
-                            error_type=fields["error_type"],
-                            error_code=fields["error_code"],
-                            exc_info=exc_info,
-                            path=fields["path"],
-                        )
-                    elif status_code == 429:
-                        logger.warning(
-                            "http.request.rate_limited",
-                            method=fields["method"],
-                            route=fields["route"],
-                            status_code=fields["status_code"],
-                            duration_ms=fields["duration_ms"],
-                        )
-                    else:
-                        streaming = content_type.startswith("text/event-stream")
-                        elapsed = (
-                            (response_started_at - started_at)
-                            if streaming and response_started_at is not None
-                            else (finished_at - started_at)
-                        )
-                        if (
-                            elapsed >= self.SLOW_SECONDS
-                            and fields["route"] not in self.HELD_ROUTES
-                        ):
-                            fields["duration_ms"] = round(elapsed * 1000, 1)
-                            fields["latency_kind"] = (
-                                "time_to_first_byte" if streaming else "total"
-                            )
-                            logger.warning(
-                                "http.request.slow",
-                                method=fields["method"],
-                                route=fields["route"],
-                                status_code=fields["status_code"],
-                                duration_ms=fields["duration_ms"],
-                                latency_kind=fields["latency_kind"],
-                                path=fields["path"],
-                            )
-                        elif settings.local_http_access_logs_enabled:
-                            logger.info(
-                                "http.request.local_completed",
-                                method=fields["method"],
-                                route=fields["route"],
-                                status_code=fields["status_code"],
-                                duration_ms=fields["duration_ms"],
-                            )
-                        else:
-                            logger.debug(
-                                "http.request.completed",
-                                method=fields["method"],
-                                route=fields["route"],
-                                status_code=fields["status_code"],
-                                duration_ms=fields["duration_ms"],
-                            )
-
-    @staticmethod
-    def _route_template(scope: dict) -> str:
-        route = scope.get("route")
-        value = getattr(route, "path_format", None) or getattr(route, "path", None)
-        return value if isinstance(value, str) else _UNMATCHED_ROUTE
-
-
-# Compatibility name retained for imports and generated SDK tests.
-RequestIdMiddleware = RequestObserverMiddleware
-
-
-class RequestBodyLimitMiddleware:
-    """Enforce a byte ceiling without trusting the Content-Length header."""
-
-    def __init__(self, app, max_bytes: int):
-        self.app = app
-        self.max_bytes = max_bytes
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] != "http" or self.max_bytes <= 0:
-            await self.app(scope, receive, send)
-            return
-
-        headers = dict(scope.get("headers") or [])
-        request_id = headers.get(b"x-request-id", b"").decode("latin-1") or None
-        content_length = headers.get(b"content-length")
-        if content_length is not None:
-            try:
-                if int(content_length) > self.max_bytes:
-                    await self._send_too_large(scope, receive, send, request_id)
-                    return
-            except ValueError:
-                pass
-
-        received = 0
-
-        async def receive_limited():
-            nonlocal received
-            message = await receive()
-            if message["type"] == "http.request":
-                received += len(message.get("body", b""))
-                if received > self.max_bytes:
-                    raise PayloadTooLargeError(max_bytes=self.max_bytes)
-            return message
-
-        try:
-            await self.app(scope, receive_limited, send)
-        except PayloadTooLargeError:
-            await self._send_too_large(scope, receive, send, request_id)
-
-    async def _send_too_large(self, scope, receive, send, request_id):
-        response = JSONResponse(
-            status_code=413,
-            content={
-                "message": "request exceeds the maximum allowed size",
-                "code": "UPLOAD_TOO_LARGE",
-                "request_id": request_id,
-                "details": {"field": "request", "max_bytes": self.max_bytes},
-            },
-        )
-        await response(scope, receive, send)
 
 
 def create_app(modules=OSS_MODULES) -> FastAPI:
@@ -735,205 +426,11 @@ def create_app(modules=OSS_MODULES) -> FastAPI:
     # Routers — registered from the module registry (app/core/registry).
     # Order follows OSS_MODULES; intra-module order follows each module's
     # routers() thunk. See app/modules/<name>/module.py.
+    # Before the modules: probes must answer even if a module's router raises
+    # while being included, which is exactly when a readiness probe is worth
+    # having.
+    app.include_router(health_router)
     include_module_routers(app, modules)
-
-    # Liveness: process/event-loop check only. No DB or network dependency, so
-    # it normally completes within ~100 ms. 503 when the event loop is wedged
-    # (lag over the unhealthy threshold), so a liveness probe restarts the
-    # process. A fully blocked loop can't serve this at all, which trips the
-    # probe's timeout — either way a hung process is restarted instead of
-    # hanging silently.
-    @app.get("/health/live", include_in_schema=False)
-    @app.get("/livez", include_in_schema=False)
-    async def health_live():
-        from app.core.observability.loop_watchdog import (
-            get_loop_lag_seconds,
-            is_loop_healthy,
-        )
-
-        healthy = is_loop_healthy()
-        payload = {
-            "status": "ok" if healthy else "unhealthy",
-            "loop_lag_seconds": round(get_loop_lag_seconds(), 3),
-        }
-        return JSONResponse(payload, status_code=200 if healthy else 503)
-
-    #: One degraded/recovered pair per dependency instead of a record per
-    #: probe. Threshold 1: readiness is asked constantly, so the first failure
-    #: is already the transition worth reporting.
-    _readiness_incidents = {
-        name: DependencyIncident(name, logger=logger, degradation_threshold=1)
-        for name in ("db", "redis", "supertokens")
-    }
-
-    # Readiness: bounded, concurrent checks for dependencies required to serve
-    # new work. Each check has ~1 s; the whole endpoint has a ~2 s deadline.
-    # 503 when not ready; only generic component states are exposed, never
-    # connection strings or provider responses.
-    @app.get("/health/ready", include_in_schema=False)
-    async def health_ready():
-        import asyncio as _asyncio
-
-        async def _probe(name: str, check) -> bool:
-            """One bounded dependency check that says why it failed, once.
-
-            This used to be three silent `except Exception: return False`. The
-            endpoint then reported `"db": "down"` with no reason anywhere --
-            and a prober asks every few seconds, so the obvious repair, a
-            record per attempt, is a wall of identical lines during exactly the
-            outage someone is trying to read. `DependencyIncident` emits one
-            degraded record when it starts failing and one when it recovers.
-            """
-            incident = _readiness_incidents[name]
-            try:
-                healthy = bool(await _asyncio.wait_for(check(), timeout=1.0))
-            except Exception as exc:
-                incident.record_failure(error_type=type(exc).__name__)
-                return False
-            if healthy:
-                incident.record_success()
-            else:
-                incident.record_failure(error_type="unavailable")
-            return healthy
-
-        # Concurrent, each individually bounded and the set bounded by the
-        # gather. SuperTokens is in here because `initialize_supertokens` makes
-        # no network call while `verify_auth` calls the core on every
-        # authenticated request: its outage leaves readiness at 200 and the
-        # whole product unusable.
-        tasks: list[_asyncio.Task[object]] = []
-        try:
-            embedded = getattr(app.state, "embedded_worker", False)
-            tasks = [
-                create_inherited_task(_probe("db", database_reachable)),
-                create_inherited_task(_probe("redis", channel_service.ping)),
-                create_inherited_task(
-                    _probe("supertokens", supertokens_core_reachable)
-                ),
-                create_inherited_task(worker_readiness_state(embedded=embedded)),
-                create_inherited_task(schema_migration_state()),
-            ]
-            db_ok, redis_ok, auth_ok, worker, schema = await _asyncio.wait_for(
-                _asyncio.gather(*tasks), timeout=2.0
-            )
-        except Exception:
-            # Readiness itself failing to run is not "the database is down";
-            # it answers 503 either way, so the log is the only place the
-            # difference can be seen.
-            logger.error("app.health_ready.probe_failed.failed", exc_info=True)
-            db_ok, redis_ok, auth_ok, worker, schema = False, False, False, None, None
-        finally:
-            for task in tasks:
-                if not task.done():
-                    task.cancel()
-
-        report = readiness.build_readiness_report(
-            components={
-                "db": readiness.dependency_state(db_ok),
-                "redis": readiness.dependency_state(redis_ok),
-                "supertokens": readiness.dependency_state(auth_ok),
-                "worker": readiness.worker_state(worker),
-                "migrations": readiness.migrations_state(schema) if schema else None,
-            },
-            instance_id=settings.lemma_runtime_instance_id,
-        )
-        return JSONResponse(report.payload, status_code=report.status_code)
-
-    @app.get("/health/capabilities", include_in_schema=False)
-    async def health_capabilities():
-        from app.modules.datastore.module import embedding_capability
-        from app.modules.pod_bundle.config import pod_bundle_settings
-
-        embeddings = embedding_capability()
-        payload = {
-            "status": (
-                "degraded" if embeddings.status == "degraded" else embeddings.status
-            ),
-            "capabilities": {
-                "embeddings": {
-                    "status": embeddings.status,
-                    "detail": embeddings.detail,
-                }
-            },
-        }
-        payload["capabilities"]["sandbox"] = sandbox_capability()
-        if settings.lemma_local_ai_ready is not None:
-            payload["capabilities"]["ai_profile"] = {
-                "status": "ready" if settings.lemma_local_ai_ready else "needs_setup",
-                "detail": (
-                    "Local AI provider is configured"
-                    if settings.lemma_local_ai_ready
-                    else "Configure an AI provider in Lemma Control Center"
-                ),
-            }
-        # What this deployment is configured to *do*, for a client that has to
-        # decide whether a behaviour it depends on is reachable here at all. The
-        # product scenario suite reads this rather than a local `.env` file:
-        # pointed at a deployment, a local file describes a different machine,
-        # so a suite trusting it skips and runs for the wrong reasons.
-        #
-        # `environment` and `llm_mode` are reported everywhere. A deployment
-        # serving the scripted test model is misconfigured, and that is worth
-        # being visible rather than hidden.
-        #
-        # The rest is this deployment's security posture — whether signup is
-        # rate limited, whether a connector may reach a private address — and
-        # the honest answer to a stranger asking "are your gates on?" is that it
-        # is none of their business. Written as "not production", that was one
-        # environment value narrower than the principle: a staging or preview
-        # deployment on the internet runs as `development` and advertised which
-        # of its abuse controls were off. The block is for the scenario suite,
-        # which runs locally.
-        configuration: dict[str, object] = {
-            "environment": settings.environment,
-            "llm_mode": settings.e2e_llm_mode,
-        }
-        if settings.is_local_mode():
-            configuration |= {
-                "abuse_protection": settings.auth_abuse_protection_enabled,
-                "altcha": settings.auth_altcha_enabled,
-                "email_verification_required": (
-                    settings.auth_email_verification_required
-                ),
-                "email_deliverability_checks": (
-                    settings.auth_email_deliverability_checks_enabled
-                ),
-                "disposable_email_domains": (
-                    settings.auth_disposable_email_domains_enabled
-                ),
-                "private_network_targets": (
-                    settings.connector_allow_private_network_targets
-                ),
-                "role_cache_ttl_seconds": (
-                    settings.authorization_role_cache_ttl_seconds
-                ),
-                "bundle_daily_export_limit": (
-                    pod_bundle_settings.pod_bundle_daily_export_limit
-                ),
-                "bundle_daily_import_limit": (
-                    pod_bundle_settings.pod_bundle_daily_import_limit
-                ),
-                "usage_limit_overrides": bool(settings.usage_org_limit_overrides_json),
-            }
-        payload["configuration"] = configuration
-        if settings.lemma_runtime_instance_id:
-            payload["instance_id"] = settings.lemma_runtime_instance_id
-        return payload
-
-    # Compatibility alias for /health/live during probe migration.
-    @app.get("/health", include_in_schema=False)
-    async def health_alias():
-        from app.core.observability.loop_watchdog import (
-            get_loop_lag_seconds,
-            is_loop_healthy,
-        )
-
-        healthy = is_loop_healthy()
-        payload = {
-            "status": "ok" if healthy else "unhealthy",
-            "loop_lag_seconds": round(get_loop_lag_seconds(), 3),
-        }
-        return JSONResponse(payload, status_code=200 if healthy else 503)
 
     # Registered only alongside the document it renders. Left on with
     # `openapi_url=None` it would serve a reference UI pointed at nothing.
