+++ lemma-backend/app/core/authorization/service.py
@@ -6,20 +6,17 @@ from dataclasses import dataclass
 from datetime import datetime
 from uuid import UUID
 
-from sqlalchemy import delete, exists, select
+from sqlalchemy import delete, select
 from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.bounded import BoundedSet
 from app.core.authorization.context import (
     ActorType,
-    AuthorizationDecision,
     Context,
     PrincipalRef,
     ResourceRef,
     ResourceType,
-    ResourceVisibility,
-    normalize_resource_visibility,
 )
 from app.core.authorization.cache import (
     RoleSnapshot,
@@ -28,23 +25,17 @@ from app.core.authorization.cache import (
     set_role_snapshot,
 )
 from app.core.authorization.delegation import (
-    DESTRUCTIVE_ACTIONS,
     DelegationClaims,
-    WorkloadPrincipalType,
 )
-from app.core.authorization.session_approvals import has_session_approval
 
 
 from app.core.domain.errors import DomainError
 from app.core.infrastructure.db.transaction_locks import connection_released
 from app.core.authorization.grants import (
-    HUMAN_GRANTEE_TYPES,
     delete_grantee_grants,
-    grant_resource_type_values,
 )
 from app.core.authorization.models import (
     AuthPermissionModel,
-    ResourcePermissionGrantModel,
     RoleAssignmentModel,
     RoleModel,
     RolePermissionModel,
@@ -53,59 +44,21 @@ from app.core.authorization.permissions import (
     PERMISSION_BY_ID,
     PERMISSION_DEFINITIONS,
     SYSTEM_ROLE_PERMISSIONS,
-    PermissionScope,
-    Permissions,
-    equivalent_permission_ids,
 )
-from app.core.authorization.resource_actions import owner_actions_for_resource
-from app.core.authorization.workload_authority import authorize_delegated_workload
+from app.core.authorization.authorizer import Authorizer
 from app.core.authorization.resource_names import resolve_resource_id_by_name
 from app.core.authorization.resource_tables import (
     RESOURCE_TABLES,
-    pod_is_unknowable,
 )
 from app.modules.identity.infrastructure.models.organization_models import (
     OrganizationMember,
 )
-from app.modules.connectors.infrastructure.models.account import Account
-from app.modules.connectors.infrastructure.models.auth_config import AuthConfig
 from app.core.authorization.roles import (
     SYSTEM_POD_ROLE_NAMES,
     normalize_role_list,
     normalize_role_name,
 )
 from app.modules.pod.infrastructure.models.pod_models import Pod, PodMember
-
-
-async def _session_approval(
-    ctx: "Context",
-    *,
-    session_id: str | None,
-    workload_actor_id: str | None,
-    permission_id: str,
-) -> bool:
-    """``has_session_approval``, asked at most once per permission per request.
-
-    The lookup is a Redis round trip made while the request's pooled database
-    connection is checked out, so repeating it per check is the expensive part.
-
-    Caching it means an approval granted *during* a request is not seen by that
-    request. That is correct for an ordinary request, which is short; for a
-    long-lived streamed one it means the approval takes effect on the next call
-    rather than mid-stream. Acceptable because approvals are awaited through the
-    approval flow rather than polled here -- but it is a staleness window, not
-    an invariant.
-    """
-    cached = ctx._session_approval_cache.get(permission_id)  # noqa: SLF001
-    if cached is not None:
-        return cached
-    approved = await has_session_approval(
-        session_id=session_id,
-        workload_actor_id=workload_actor_id,
-        permission_id=permission_id,
-    )
-    ctx._session_approval_cache[permission_id] = approved  # noqa: SLF001
-    return approved
 
 
 SYSTEM_ORG_ROLES = {"ORG_MEMBER", "ORG_EDITOR", "ORG_OWNER"}
@@ -1112,815 +1065,3 @@ class AuthorizationDataService:
             principal_id=targets.organization_member_id,
         )
         await self._invalidate_snapshots_after_commit(user_id=targets.user_id)
-
-
-#: Datastore's file columns, via the one table that names them. `FOLDER` and
-#: `DOCUMENT` share a row, so either serves.
-_FILES = RESOURCE_TABLES[ResourceType.FOLDER]
-
-
-class Authorizer:
-    def __init__(self, session: AsyncSession):
-        self.session = session
-        # Grant rows for one (pod, resource-target set, principal group), keyed
-        # without the permission. See `_grant_rows_for_principal_group`.
-        self._grant_rows: dict[
-            tuple[UUID, tuple[str, ...], tuple[UUID, ...], frozenset[PrincipalRef]],
-            list[tuple[UUID, str]],
-        ] = {}
-        # Ancestor-folder ids by (pod, candidate paths). See
-        # `_acceptable_grant_resource_ids`.
-        self._folder_ids_by_paths: dict[tuple[UUID, tuple[str, ...]], list[UUID]] = {}
-
-    async def _describe_resource(self, resource: ResourceRef | None) -> str | None:
-        """Human name for a denied resource, or None when it can't be resolved.
-
-        Runs only on the denial path, so the extra lookup costs nothing in the
-        allow case. Folders and documents already carry their path; everything
-        else resolves id -> name through the same registry the grant APIs use.
-        """
-        if resource is None:
-            return None
-        if resource.path:
-            return resource.path
-        if resource.resource_id is None or resource.pod_id is None:
-            return None
-        try:
-            from app.core.authorization.resource_names import (
-                resolve_resource_names_by_ids,
-            )
-
-            resolved = await resolve_resource_names_by_ids(
-                self.session,
-                pod_id=resource.pod_id,
-                refs=[(resource.resource_type, resource.resource_id)],
-            )
-        except Exception:  # noqa: BLE001 — a nicer message is never worth an error
-            return None
-        return resolved.get((resource.resource_type, resource.resource_id))
-
-    async def authorize(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef | None = None,
-    ) -> AuthorizationDecision:
-        if permission_id not in PERMISSION_BY_ID:
-            return AuthorizationDecision(
-                False, "UNKNOWN_PERMISSION", permission_id, resource
-            )
-        if ctx.is_superuser:
-            return AuthorizationDecision(True, "SUPERUSER", permission_id, resource)
-        if ctx.actor_type == ActorType.ANONYMOUS:
-            if resource and await self._is_public_read(permission_id, resource):
-                return AuthorizationDecision(
-                    True, "PUBLIC_RESOURCE", permission_id, resource
-                )
-            return AuthorizationDecision(
-                False, "AUTH_REQUIRED", permission_id, resource
-            )
-
-        # The default pod agent mirrors the invoking user's authority but ONLY
-        # within its pinned pod: it may exercise any pod-scoped action the user
-        # holds here, while org-scoped actions and other pods are denied at this
-        # layer and must go through the user-approval-gated tools instead. Named
-        # agent/function workloads are not user-equivalent and take the
-        # stricter intersection path in ``workload_authority`` below.
-        clamp_to_pod = (
-            ctx.actor_type == ActorType.DELEGATED_USER_WORKLOAD
-            and ctx.is_user_equivalent
-        )
-        if clamp_to_pod and not self._is_pod_scoped_permission(permission_id):
-            return AuthorizationDecision(
-                False, "DELEGATED_POD_SCOPE_ONLY", permission_id, resource
-            )
-
-        if resource is None:
-            # Destructive capability checks (no specific resource) must still be
-            # gated for delegated workloads, before the user-mirror allows them.
-            destructive = await self._destructive_delegated_decision(
-                ctx, permission_id, None
-            )
-            if destructive is not None:
-                return destructive
-            if not ctx.has_permission(permission_id):
-                return AuthorizationDecision(
-                    False,
-                    "INSUFFICIENT_PERMISSION",
-                    permission_id,
-                    resource,
-                )
-            return AuthorizationDecision(
-                True, "PERMISSION_MATCH", permission_id, resource
-            )
-
-        hydrated = await self._hydrate_resource(resource)
-        if clamp_to_pod and (
-            hydrated.pod_id != ctx.pod_id
-            if hydrated.pod_id is not None
-            # A pod-scoped resource whose pod could not be established is
-            # refused rather than waved through. This used to read
-            # `pod_id is not None and pod_id != ctx.pod_id`, so a type
-            # hydration did not know skipped the clamp entirely — the guard
-            # confining a pod's default agent to its own pod was opt-out by
-            # omission. `_pod_is_unknowable` exempts the types that genuinely
-            # have no pod, and `_assert_every_resource_type_is_classified`
-            # makes staying silent about a new one impossible.
-            else pod_is_unknowable(hydrated)
-        ):
-            return AuthorizationDecision(
-                False, "DELEGATED_POD_SCOPE_ONLY", permission_id, hydrated
-            )
-        # Destructive actions are gated for EVERY delegated workload — default
-        # pod agent and named workloads alike — and this runs before the owner /
-        # org-owner / function-self shortcuts so none of them can bypass it. It
-        # returns None (proceed) only when the user approved the action for the
-        # session or a named workload holds an explicit grant (standing
-        # authority); otherwise it denies.
-        destructive = await self._destructive_delegated_decision(
-            ctx, permission_id, hydrated
-        )
-        if destructive is not None:
-            return destructive
-        if self._is_function_self_read(ctx, permission_id, hydrated):
-            return AuthorizationDecision(
-                True, "FUNCTION_SELF_READ", permission_id, hydrated
-            )
-        if self._is_org_owner_of_pod(ctx, permission_id, hydrated):
-            return AuthorizationDecision(True, "ORG_OWNER_POD", permission_id, hydrated)
-        if (
-            ctx.actor_type == ActorType.DELEGATED_USER_WORKLOAD
-            and ctx.workload_principal_refs
-        ):
-            return await authorize_delegated_workload(
-                self,
-                ctx,
-                permission_id,
-                hydrated,
-            )
-        if (
-            hydrated.owner_user_id is not None
-            and hydrated.owner_user_id == ctx.user_id
-            and permission_id in owner_actions_for_resource(hydrated.resource_type)
-        ):
-            return AuthorizationDecision(
-                True, "RESOURCE_OWNER", permission_id, hydrated
-            )
-        visibility_decision = self._visibility_read_decision(
-            ctx, permission_id, hydrated
-        )
-        if visibility_decision is not None:
-            return visibility_decision
-        if not ctx.has_permission(permission_id):
-            grant_decision = await self._resource_grant_decision(
-                ctx,
-                permission_id,
-                hydrated,
-            )
-            if grant_decision is not None:
-                return grant_decision
-            return AuthorizationDecision(
-                False,
-                "INSUFFICIENT_PERMISSION",
-                permission_id,
-                hydrated,
-            )
-        if hydrated.organization_id is not None and hydrated.pod_id is None:
-            if hydrated.organization_id != ctx.organization_id:
-                return AuthorizationDecision(
-                    False,
-                    "ORG_SCOPE_MISMATCH",
-                    permission_id,
-                    hydrated,
-                )
-            return AuthorizationDecision(True, "ORG_VISIBLE", permission_id, hydrated)
-
-        visibility = hydrated.visibility or ResourceVisibility.POD
-        if visibility == ResourceVisibility.PUBLIC:
-            return AuthorizationDecision(
-                True, "PUBLIC_RESOURCE", permission_id, hydrated
-            )
-        if hydrated.owner_user_id is not None and hydrated.owner_user_id == ctx.user_id:
-            return AuthorizationDecision(
-                True, "RESOURCE_OWNER", permission_id, hydrated
-            )
-        if visibility == ResourceVisibility.PERSONAL:
-            return AuthorizationDecision(
-                False, "PERSONAL_RESOURCE_DENIED", permission_id, hydrated
-            )
-        if visibility == ResourceVisibility.POD:
-            if hydrated.pod_id is not None and hydrated.pod_id != ctx.pod_id:
-                return AuthorizationDecision(
-                    False, "POD_SCOPE_MISMATCH", permission_id, hydrated
-                )
-            return AuthorizationDecision(True, "POD_VISIBLE", permission_id, hydrated)
-        if visibility == ResourceVisibility.RESTRICTED:
-            grant_decision = await self._resource_grant_decision(
-                ctx,
-                permission_id,
-                hydrated,
-            )
-            if grant_decision is not None:
-                return grant_decision
-            return AuthorizationDecision(
-                False, "MISSING_RESOURCE_GRANT", permission_id, hydrated
-            )
-        return AuthorizationDecision(
-            False, "UNSUPPORTED_VISIBILITY", permission_id, hydrated
-        )
-
-    async def _destructive_delegated_decision(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef | None,
-    ) -> AuthorizationDecision | None:
-        """Gate DESTRUCTIVE_ACTIONS for delegated workloads.
-
-        Returns a deny decision when the action must be blocked, or ``None`` to
-        let normal authorization proceed. Proceed only when the user recorded a
-        session approval (``APPROVE_FOR_SESSION``) for the action type, or a
-        named workload holds an explicit grant for it (standing authority — the
-        headless path). Applies to the default pod agent and named workloads
-        alike, and is called before the owner / org-owner / function-self
-        shortcuts so those cannot bypass it.
-
-        ``None`` here is only a pass through *this* gate: a named workload then
-        still has to clear the invoker ceiling in ``workload_authority``, so an
-        approval unlocks a destructive action without conferring authority the
-        approving person does not have.
-        """
-        if (
-            ctx.actor_type != ActorType.DELEGATED_USER_WORKLOAD
-            or permission_id not in DESTRUCTIVE_ACTIONS
-        ):
-            return None
-        if await _session_approval(
-            ctx,
-            session_id=ctx.delegation_session_id,
-            workload_actor_id=ctx.actor_id,
-            permission_id=permission_id,
-        ):
-            return None
-        if resource is not None and ctx.workload_principal_refs:
-            grant_ids = await self._matching_grant_ids_for_principal_sets(
-                ctx, permission_id, resource, (ctx.workload_principal_refs,)
-            )
-            if grant_ids:
-                return None
-        return AuthorizationDecision(
-            False, "DESTRUCTIVE_ACTION_REQUIRES_APPROVAL", permission_id, resource
-        )
-
-    @staticmethod
-    def _is_read_permission(permission_id: str) -> bool:
-        return permission_id.endswith(".read")
-
-    def _visibility_read_decision(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef,
-    ) -> AuthorizationDecision | None:
-        """Allow a read the resource's own visibility already permits.
-
-        Visibility above POD is a property of the resource, not of the viewer's
-        role, so it must be evaluated before the pod-permission gate below. It
-        used to sit *after* that gate, which made PUBLIC unreachable: a non-member
-        holds no pod permissions, failed the gate, and was denied before the
-        PUBLIC branch ever ran. The level changed nothing for members (who match
-        POD anyway) and granted nothing to anyone else — while ANONYMOUS, checked
-        earlier in ``authorize``, did get through. Signed-out beat signed-in.
-
-        Deliberately narrow: reads only, humans only. Write-shaped actions and
-        every workload actor fall through to the paths they already took, so
-        flipping a resource to PUBLIC can never widen what an agent or function
-        may do, nor let anyone edit what they can now read.
-        """
-        if ctx.actor_type != ActorType.USER or ctx.user_id is None:
-            return None
-        if not self._is_read_permission(permission_id):
-            return None
-        if resource.pod_id is not None and resource.pod_id != ctx.pod_id:
-            return None
-
-        visibility = resource.visibility or ResourceVisibility.POD
-        if visibility == ResourceVisibility.PUBLIC:
-            return AuthorizationDecision(
-                True, "PUBLIC_RESOURCE", permission_id, resource
-            )
-        return None
-
-    @staticmethod
-    def _is_pod_scoped_permission(permission_id: str) -> bool:
-        definition = PERMISSION_BY_ID.get(permission_id)
-        return definition is not None and definition.scope == PermissionScope.POD
-
-    @staticmethod
-    def _is_org_owner_of_pod(
-        ctx: Context, permission_id: str, resource: ResourceRef
-    ) -> bool:
-        """Org owners have full authority over a pod and everything inside it.
-
-        The pod read/list/delete services already treat the organization owner as
-        able to manage any pod in their organization. This mirrors that intent at
-        the authorization layer for the pod entity AND its pod-scoped child
-        resources (apps, agents, functions, ...), so an org owner whose pod
-        authority comes only from the org-owner shortcut can create/deploy/update
-        them instead of hitting a 403 the service would have allowed (e.g. the
-        app bundle upload that follows app.create). Org-level actions are not
-        covered: they still require the org permission itself.
-        """
-        if ctx.actor_type != ActorType.USER and not ctx.is_user_equivalent:
-            return False
-        if not Authorizer._is_pod_scoped_permission(permission_id):
-            return False
-        if ctx.pod_id is None or resource.pod_id != ctx.pod_id:
-            return False
-        return "ORG_OWNER" in ctx.role_names
-
-    @staticmethod
-    def _is_function_self_read(
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef,
-    ) -> bool:
-        if permission_id != Permissions.FUNCTION_READ:
-            return False
-        if (
-            resource.resource_type != ResourceType.FUNCTION
-            or resource.resource_id is None
-        ):
-            return False
-        if ctx.actor_type == ActorType.FUNCTION and ctx.actor_id == str(
-            resource.resource_id
-        ):
-            return True
-        if ctx.actor_type != ActorType.DELEGATED_USER_WORKLOAD:
-            return False
-        return PrincipalRef(
-            WorkloadPrincipalType.FUNCTION.value.upper(), resource.resource_id
-        ) in (ctx.workload_principal_refs)
-
-    async def accessible_resource_ids(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource_type: ResourceType,
-        pod_id: UUID,
-    ) -> frozenset[UUID]:
-        principal_sets = ctx.grant_principal_sets or (ctx.principal_refs,)
-        if (
-            ctx.actor_type == ActorType.DELEGATED_USER_WORKLOAD
-            and ctx.workload_principal_refs
-        ):
-            # The workload's own grants: the FIRST half of the intersection
-            # ``workload_authority`` applies. The second half — could
-            # the invoking person reach it themselves — is not expressible as a
-            # grant query, because their access also comes from role
-            # permissions, ownership and visibility. So this narrows to what
-            # the workload was granted and nothing more; a caller listing for a
-            # delegated workload must still authorize each id it shows, which
-            # is where the person's ceiling is applied. Also collapses the
-            # per-principal-group query loop to a single query here.
-            principal_sets = (ctx.workload_principal_refs,)
-        if not principal_sets or any(not group for group in principal_sets):
-            return frozenset()
-        permission_ids = equivalent_permission_ids(permission_id)
-        from sqlalchemy import and_, or_
-
-        visible_ids: set[UUID] | None = None
-        for principal_group in principal_sets:
-            principal_clauses = [
-                and_(
-                    ResourcePermissionGrantModel.grantee_type == principal.type,
-                    ResourcePermissionGrantModel.grantee_id == principal.id,
-                )
-                for principal in principal_group
-            ]
-            resource_type_values = grant_resource_type_values(resource_type)
-            stmt = select(ResourcePermissionGrantModel.resource_id).where(
-                ResourcePermissionGrantModel.pod_id == pod_id,
-                ResourcePermissionGrantModel.resource_type.in_(resource_type_values),
-                ResourcePermissionGrantModel.permission_id.in_(permission_ids),
-                or_(*principal_clauses),
-            )
-            group_ids = set((await self.session.execute(stmt)).scalars().all())
-            visible_ids = group_ids if visible_ids is None else visible_ids & group_ids
-        return frozenset(visible_ids or set())
-
-    async def hydrate_resource_ref(self, resource: ResourceRef) -> ResourceRef:
-        """Fill in visibility / owner / path for a bare ``ResourceRef``.
-
-        ``resolve_resource_ref`` returns only the identity triple; ``authorize``
-        hydrates internally. Callers that need to *report* on a resource rather
-        than just authorize it (the share preview) need the same fields, and
-        should not reach into the private helper to get them.
-        """
-        return await self._hydrate_resource(resource)
-
-    async def _is_public_read(self, permission_id: str, resource: ResourceRef) -> bool:
-        if not permission_id.endswith(".read"):
-            return False
-        hydrated = await self._hydrate_resource(resource)
-        return hydrated.visibility == ResourceVisibility.PUBLIC
-
-    async def _hydrate_resource(self, resource: ResourceRef) -> ResourceRef:
-        if resource.visibility is not None:
-            return resource
-        if resource.resource_id is None:
-            return resource
-        # Folder/document hydration also fetches the row's path so folder grants
-        # can cascade to descendants in the matcher.
-        if resource.resource_type is ResourceType.POD:
-            # A pod's own id is the pod it belongs to, so this needs no table.
-            # Every caller happens to pass `pod_id` today; deriving it here is
-            # what stops the clamp depending on that continuing to be true.
-            return ResourceRef(
-                resource_type=resource.resource_type,
-                resource_id=resource.resource_id,
-                organization_id=resource.organization_id,
-                pod_id=resource.pod_id or resource.resource_id,
-                owner_user_id=resource.owner_user_id,
-                visibility=resource.visibility,
-            )
-        if resource.resource_type in (ResourceType.FOLDER, ResourceType.DOCUMENT):
-            return await self._hydrate_datastore_file(resource)
-        if resource.resource_type == ResourceType.CONNECTOR:
-            return await self._hydrate_connector(resource)
-        if resource.resource_type == ResourceType.CONNECTOR_ACCOUNT:
-            return await self._hydrate_connector_account(resource)
-        if resource.resource_type == ResourceType.CONNECTOR_AUTH_CONFIG:
-            return await self._hydrate_connector_auth_config(resource)
-        table = RESOURCE_TABLES.get(resource.resource_type)
-        if table is None:
-            # Not a silent pass-through: every type is classified below, and
-            # `_assert_every_resource_type_is_classified` refuses to import this
-            # module if one is not. Reaching here means a type is registered as
-            # unroutable and something built a ref for it anyway.
-            return resource
-        if table.visibility_column is None:
-            stmt = select(table.pod_column, table.owner_column).where(
-                table.id_column == resource.resource_id
-            )
-        else:
-            stmt = select(
-                table.pod_column, table.owner_column, table.visibility_column
-            ).where(table.id_column == resource.resource_id)
-        row = (await self.session.execute(stmt)).first()
-        if row is None:
-            return resource
-        if table.visibility_column is None:
-            visibility = ResourceVisibility.PERSONAL
-        else:
-            visibility = self._normalize_visibility(row[2])
-        return ResourceRef(
-            resource_type=resource.resource_type,
-            resource_id=resource.resource_id,
-            organization_id=resource.organization_id,
-            pod_id=resource.pod_id or row[0],
-            owner_user_id=resource.owner_user_id or row[1],
-            visibility=visibility,
-        )
-
-    async def _hydrate_datastore_file(self, resource: ResourceRef) -> ResourceRef:
-        """Hydrate a FOLDER/DOCUMENT ref, including its path so folder grants
-        can cascade to descendants."""
-        # A path-only check carries the POD id in resource_id — the caller uses
-        # `resource_id or pod_id` because a ResourceRef wants one, and the path
-        # is what the grant cascade actually matches on. Querying for a file row
-        # whose id equals a pod id can only ever return nothing, so skip it:
-        # this ran on every datastore path check.
-        if resource.resource_id is None or resource.resource_id == resource.pod_id:
-            return resource
-
-        stmt = select(
-            _FILES.pod_column,
-            _FILES.owner_column,
-            _FILES.visibility_column,
-            _FILES.path_column,
-        ).where(_FILES.id_column == resource.resource_id)
-        row = (await self.session.execute(stmt)).first()
-        if row is None:
-            return resource
-        return ResourceRef(
-            resource_type=resource.resource_type,
-            resource_id=resource.resource_id,
-            organization_id=resource.organization_id,
-            pod_id=resource.pod_id or row[0],
-            owner_user_id=resource.owner_user_id or row[1],
-            visibility=resource.visibility or self._normalize_visibility(row[2]),
-            path=resource.path or row[3],
-        )
-
-    async def _hydrate_connector(self, resource: ResourceRef) -> ResourceRef:
-        if resource.pod_id is None:
-            return resource
-        return ResourceRef(
-            resource_type=resource.resource_type,
-            resource_id=resource.resource_id,
-            organization_id=resource.organization_id,
-            pod_id=resource.pod_id,
-            owner_user_id=resource.owner_user_id,
-            # Connectors are org-wide capability resources, always
-            # available to everyone: a grant on the app is a capability grant
-            # ("may use this app type"), never a sharing grant, so it must not
-            # restrict the app. The real access boundary is the connected
-            # *account*, which is user-owned and enforced separately in
-            # ``account_resolution_service`` (own account is returned directly;
-            # someone else's requires ``connector_account.use``). Keeping the
-            # app POD-visible means a workload holding ``connector.use``
-            # is gated on that capability alone, regardless of which user the run
-            # is delegated for -- the regression that denied delegated runs for
-            # any member who did not also personally hold an app grant.
-            visibility=ResourceVisibility.POD,
-        )
-
-    async def _hydrate_connector_account(self, resource: ResourceRef) -> ResourceRef:
-        stmt = select(
-            Account.organization_id,
-            Account.user_id,
-        ).where(
-            Account.id == resource.resource_id,
-        )
-        row = (await self.session.execute(stmt)).first()
-        if row is None:
-            return resource
-        pod_id = resource.pod_id
-        return ResourceRef(
-            resource_type=resource.resource_type,
-            resource_id=resource.resource_id,
-            organization_id=resource.organization_id or row[0],
-            pod_id=pod_id,
-            owner_user_id=resource.owner_user_id or row[1],
-            visibility=(
-                await self._visibility_from_optional_grants(
-                    pod_id=pod_id,
-                    resource_type=resource.resource_type,
-                    resource_id=resource.resource_id,
-                )
-                if pod_id is not None
-                else resource.visibility
-            ),
-        )
-
-    async def _hydrate_connector_auth_config(
-        self, resource: ResourceRef
-    ) -> ResourceRef:
-        stmt = select(AuthConfig.organization_id, AuthConfig.created_by_user_id).where(
-            AuthConfig.id == resource.resource_id
-        )
-        row = (await self.session.execute(stmt)).first()
-        if row is None:
-            return resource
-        return ResourceRef(
-            resource_type=resource.resource_type,
-            resource_id=resource.resource_id,
-            organization_id=resource.organization_id or row[0],
-            pod_id=resource.pod_id,
-            owner_user_id=resource.owner_user_id or row[1],
-            visibility=resource.visibility,
-        )
-
-    async def _visibility_from_optional_grants(
-        self,
-        *,
-        pod_id: UUID,
-        resource_type: ResourceType,
-        resource_id: UUID,
-    ) -> ResourceVisibility:
-        conditions = [
-            ResourcePermissionGrantModel.pod_id == pod_id,
-            ResourcePermissionGrantModel.resource_type.in_(
-                grant_resource_type_values(resource_type)
-            ),
-            ResourcePermissionGrantModel.resource_id == resource_id,
-            # Human sharing grants only. A grant to an agent is a workload
-            # capability grant and says nothing about whether people may see the
-            # resource -- which is why `clear_human_sharing_grants` deletes the
-            # human ones and preserves these when a resource leaves RESTRICTED.
-            # Counting them here contradicted that: pinning a shared account on
-            # an agent was itself what made the account RESTRICTED, so the act of
-            # configuring the agent locked every pod member out of the very
-            # account it was configured to use.
-            ResourcePermissionGrantModel.grantee_type.in_(HUMAN_GRANTEE_TYPES),
-        ]
-        grant_exists_stmt = select(exists().where(*conditions))
-        has_grants = (await self.session.execute(grant_exists_stmt)).scalar_one()
-        if has_grants:
-            return ResourceVisibility.RESTRICTED
-        return ResourceVisibility.POD
-
-    async def _resource_grant_decision(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef,
-    ) -> AuthorizationDecision | None:
-        visibility = resource.visibility or ResourceVisibility.POD
-        if (
-            visibility == ResourceVisibility.PERSONAL
-            and resource.owner_user_id != ctx.user_id
-        ):
-            return None
-        grant_ids = await self._matching_grant_ids(ctx, permission_id, resource)
-        if not grant_ids:
-            return None
-        return AuthorizationDecision(
-            True,
-            "RESOURCE_GRANT_MATCH",
-            permission_id,
-            resource,
-            matched_grant_ids=tuple(grant_ids),
-        )
-
-    async def _matching_grant_ids(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef,
-    ) -> list[UUID]:
-        return await self._matching_grant_ids_for_principal_sets(
-            ctx,
-            permission_id,
-            resource,
-            ctx.grant_principal_sets or (ctx.principal_refs,),
-        )
-
-    @staticmethod
-    def _ancestor_folder_paths(path: str) -> list[str]:
-        """Return the ancestor folder paths of ``path`` (excluding itself).
-
-        ``/a/b/c.md`` -> ``["/", "/a", "/a/b"]``. The root ``/`` is included so
-        a grant on the (optional) root folder cascades pod-wide.
-        """
-        segments = [segment for segment in path.split("/") if segment]
-        paths = ["/"]
-        accumulated = ""
-        for segment in segments[:-1]:
-            accumulated = f"{accumulated}/{segment}"
-            paths.append(accumulated)
-        return paths
-
-    async def _acceptable_grant_resource_ids(
-        self, resource: ResourceRef
-    ) -> list[UUID] | None:
-        """Resource ids a grant may target to authorize ``resource``.
-
-        For hierarchical FOLDER/DOCUMENT resources this is ``{self id} ∪
-        {ancestor folder ids} ∪ {pod id (pod-wide root grant)}`` so a grant
-        on the resource itself or any ancestor folder cascades down. Returns
-        ``None`` for every other resource type, signalling the caller to keep
-        exact-id matching.
-        """
-        if resource.resource_type not in (ResourceType.FOLDER, ResourceType.DOCUMENT):
-            return None
-        if resource.pod_id is None:
-            return [resource.resource_id] if resource.resource_id else []
-        acceptable: set[UUID] = set()
-        if resource.resource_id is not None:
-            acceptable.add(resource.resource_id)
-        # A grant keyed on the pod id itself is the pod-wide document grant.
-        acceptable.add(resource.pod_id)
-        if resource.path:
-            # Resolve grant-target ids by path for this resource's OWN folder
-            # plus every ancestor folder. Including the resource's own path is
-            # what lets a grant *on* the folder authorize the folder itself when
-            # the caller authorizes by path only -- e.g. the list_files /
-            # search-scope pre-check passes resource_name but no resource_id, so
-            # ``_require_document_action`` falls back to the pod id and the
-            # self-grant would otherwise never match. This mirrors the SQL
-            # projection's self-match (``resource_path_col == granted.path``).
-            candidate_paths = [
-                *self._ancestor_folder_paths(resource.path),
-                resource.path,
-            ]
-            acceptable.update(
-                await self._folder_ids_for_paths(
-                    resource.pod_id, tuple(candidate_paths)
-                )
-            )
-        return list(acceptable)
-
-    async def _folder_ids_for_paths(
-        self, pod_id: UUID, candidate_paths: tuple[str, ...]
-    ) -> list[UUID]:
-        """Ancestor-folder ids for ``candidate_paths``, resolved once.
-
-        The paths depend on the *resource*, not on the permission being
-        checked, so a caller asking many permissions about one folder was
-        re-resolving the same ancestor chain each time. Memoized for the same
-        reason and with the same lifetime as ``_grant_rows`` below.
-        """
-        key = (pod_id, candidate_paths)
-        cached = self._folder_ids_by_paths.get(key)
-        if cached is not None:
-            return cached
-        stmt = select(_FILES.id_column).where(
-            _FILES.pod_column == pod_id,
-            _FILES.path_column.in_(candidate_paths),
-        )
-        resolved = list((await self.session.execute(stmt)).scalars().all())
-        self._folder_ids_by_paths[key] = resolved
-        return resolved
-
-    async def _matching_grant_ids_for_principal_sets(
-        self,
-        ctx: Context,
-        permission_id: str,
-        resource: ResourceRef,
-        principal_sets: tuple[frozenset[PrincipalRef], ...],
-    ) -> list[UUID]:
-        if resource.pod_id is None or resource.resource_id is None:
-            return []
-        permission_ids = equivalent_permission_ids(permission_id)
-        if not principal_sets or any(not group for group in principal_sets):
-            return []
-
-        # For folders/documents, a grant on any ancestor folder (or the pod-wide
-        # root grant) cascades down; every other resource type stays exact-match.
-        acceptable_ids = await self._acceptable_grant_resource_ids(resource)
-        if acceptable_ids is None:
-            target_ids: tuple[UUID, ...] = (resource.resource_id,)
-        elif not acceptable_ids:
-            return []
-        else:
-            target_ids = tuple(sorted(acceptable_ids))
-
-        resource_type_values = grant_resource_type_values(resource.resource_type)
-        matched_ids: list[UUID] = []
-        for principal_group in principal_sets:
-            rows = await self._grant_rows_for_principal_group(
-                pod_id=resource.pod_id,
-                resource_type_values=resource_type_values,
-                target_ids=target_ids,
-                principal_group=principal_group,
-            )
-            group_ids = [
-                grant_id
-                for grant_id, granted_permission_id in rows
-                if granted_permission_id in permission_ids
-            ]
-            if not group_ids:
-                return []
-            matched_ids.extend(group_ids)
-        return matched_ids
-
-    async def _grant_rows_for_principal_group(
-        self,
-        *,
-        pod_id: UUID,
-        resource_type_values: tuple[str, ...],
-        target_ids: tuple[UUID, ...],
-        principal_group: frozenset[PrincipalRef],
-    ) -> list[tuple[UUID, str]]:
-        """``(grant_id, permission_id)`` for one principal group, read once.
-
-        The permission is deliberately *not* in the WHERE clause, and that is
-        the whole point. Everything else in the key — the pod, the resource
-        targets, the principals — is fixed for a caller asking many questions
-        about one resource, so filtering by permission in SQL turned N
-        permissions into N round trips over the same handful of rows.
-        ``/pods/{id}/permissions/me`` asks 51 of them; a POD_VIEWER paid ~40.
-
-        Selecting the permission alongside the id and filtering in Python
-        returns exactly the rows the per-permission query returned, so every
-        decision, reason code and matched grant id is identical by
-        construction. Nothing here knows how to rebuild a decision — it only
-        avoids asking the same question again.
-
-        The memo is per ``Authorizer``, which is built per context, so its
-        lifetime matches the decision cache that already sits in front of it.
-        """
-        from sqlalchemy import or_, and_
-
-        key = (pod_id, resource_type_values, target_ids, principal_group)
-        cached = self._grant_rows.get(key)
-        if cached is not None:
-            return cached
-        clauses = [
-            (
-                ResourcePermissionGrantModel.grantee_type == principal.type,
-                ResourcePermissionGrantModel.grantee_id == principal.id,
-            )
-            for principal in principal_group
-        ]
-        stmt = select(
-            ResourcePermissionGrantModel.id,
-            ResourcePermissionGrantModel.permission_id,
-        ).where(
-            ResourcePermissionGrantModel.pod_id == pod_id,
-            ResourcePermissionGrantModel.resource_type.in_(resource_type_values),
-            ResourcePermissionGrantModel.resource_id.in_(target_ids),
-            or_(*(and_(*clause) for clause in clauses)),
-        )
-        rows = [(row[0], row[1]) for row in (await self.session.execute(stmt)).all()]
-        self._grant_rows[key] = rows
-        return rows
-
-    @staticmethod
-    def _normalize_visibility(value: str | None) -> ResourceVisibility:
-        # An unreadable value falls back to POD here (rather than raising) so a
-        # malformed row degrades to pod-scoped — the safe direction — instead of
-        # 500ing the whole request.
-        return normalize_resource_visibility(value) or ResourceVisibility.POD
