+++ lemma-backend/app/core/authorization/tests/unit/test_resource_hydration.py
@@ -28,7 +28,7 @@ from app.core.authorization.context import (
     ResourceType,
     ResourceVisibility,
 )
-from app.core.authorization.service import Authorizer
+from app.core.authorization.authorizer import Authorizer
 
 pytestmark = pytest.mark.unit
