+++ lemma-backend/architecture-baseline.json
@@ -75,7 +75,6 @@
     "workflow:max": 42,
     "workflow:total": 59
   },
-  "composition_deep_imports": {},
   "core_module_imports": {
     "core->agent": 2,
     "core->apps": 2,
@@ -98,12 +97,9 @@
     "pod_bundle->function": 1,
     "workflow->agent": 4
   },
-  "induced_module_cycles": [],
-  "module_composition_imports": {},
   "module_cycles": [],
   "oversized_files": {
-    "app/app.py": 999,
-    "app/core/authorization/service.py": 1926,
+    "app/core/authorization/service.py": 1067,
     "app/core/config.py": 1301,
     "app/core/infrastructure/jobs/streaq_runtime.py": 875,
     "app/core/log/log.py": 894,
@@ -131,8 +127,8 @@
     "agent_surfaces:max": 45,
     "agent_surfaces:total": 849,
     "app:count": 2,
-    "app:max": 3,
-    "app:total": 5,
+    "app:max": 2,
+    "app:total": 4,
     "apps:count": 3,
     "apps:max": 2,
     "apps:total": 4,
